Implemented shelf selection
Shelf selection now works and is functional across both the
search page results and the current books mainpage.

Search display improved. You will now see your results shown
by either no shelf, or the shelf it is on if it already exists
in your library.

Implemented minor loading display for when books are being fetched.
Also added a 'nothing on this shelf' display when a shelf is empty.